Move create_context_cache to public utils
