+++ libs/vertexai/langchain_google_vertexai/__init__.py
@@ -6,7 +6,6 @@ from google.cloud.aiplatform_v1beta1.types import (
     Type,
 )
 
-from langchain_google_vertexai._content_catching import create_context_cache
 from langchain_google_vertexai._enums import HarmBlockThreshold, HarmCategory
 from langchain_google_vertexai.chains import create_structured_runnable
 from langchain_google_vertexai.chat_models import ChatVertexAI
@@ -28,6 +27,7 @@ from langchain_google_vertexai.gemma import (
 )
 from langchain_google_vertexai.llms import VertexAI
 from langchain_google_vertexai.model_garden import VertexAIModelGarden
+from langchain_google_vertexai.utils import create_context_cache
 from langchain_google_vertexai.vectorstores import (
     DataStoreDocumentStorage,
     GCSDocumentStorage,
