+++ libs/vertexai/langchain_google_vertexai/utils.py
